requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
dictdiffer==0.9.0
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
multidict==6.0.5
nest-asyncio==1.6.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
pytest-env==1.1.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
